Extract tap_assoc's logic into AssociactionCallTracker class
